backend/requirements.txt
fastapi>=0.109.0
uvicorn
sqlalchemy
pydantic
psycopg2-binary
pymysql
cryptography
python-dotenv
vanna>=2.0.0
openai
chromadb
faker
pandas
redis>=5.0.0
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
python-multipart
